requirements.txt
google-genai==0.2.2
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.11
openai==1.12.0
